requirements.txt
tensorflow==2.4.1
absl-py==0.11.0
numpy==1.19.5
matplotlib==3.3.4
scikit-learn==0.24.2
pandas==1.1.5
pyyaml
opencv-python